Split fat repository files into sub-domain packages (#265)
- warehouse_repository.py (1,931 lines) -> app/repositories/warehouse/:
  receiving, locations, inventory, pull_requests, audit, progress
- stock_repository.py (1,361 lines) -> app/repositories/stock/:
  items, movements, deficiency, common
- package __init__ re-exports the full public API; callers import
  'warehouse as warehouse_repository' / 'stock as stock_repository'
- pure moves, function parity verified against the old modules
