+++ backend/app/repositories/import_repository.py
@@ -694,7 +694,7 @@ def finalize_import_session(
         # Gate 1 (#224): hard inventory-sufficiency check before minting the PR. No partial pulls -
         # if any opening item can't be fully covered from available inventory, refuse creation. The
         # resolver rolls the whole finalize back (no PR, no partial state) and notifies the PO.
-        from app.repositories import warehouse_repository
+        from app.repositories import warehouse as warehouse_repository
         from app.services import notification_service
 
         needs = [
